Simpler approach to get helpers included in binary
